api/requirements.txt
pylama
pytest
pytest-cov

numpydoc==0.6.0
# June 15 2017 (artyom) : https://github.com/pyinstaller/pyinstaller/issues/2434
git+https://github.com/pyinstaller/pyinstaller.git@7e814646474efc868206beb8fb9dfe45af527109
Sphinx==1.4.8
twine==1.8.1
dill==0.2.5
Flask==0.10.1
Flask-Cors==3.0.2
Flask-SocketIO==2.5
requests==2.9.1
psutil==4.3.1
greenlet==0.4.12
gevent==1.2.2
RPi.GPIO==0.6.3

